+++ model/PersistanceLayer/class_registeredDAO.php
@@ -286,22 +286,24 @@ class registeredDAO {
 	}
 
 	/* Metodo para eliminar el usuario registrado */
-	public function deleteRegisteredUser($id) {
-
+	public function deleteRegistered($id) {
+		$proces = -1;
 		try {
-			
-			$query = ("DELETE FROM registered WHERE ID_Registered = '$id'");
-
 			$db = unserialize($_SESSION['dbconnection']);
-			$resultat = $db->getLink()->prepare($query);
-        	$resultat->execute();
-
+			$sql = ("SELECT ID_Registered FROM registered WHERE ID_Registered='$id'");
+			$resultat = $db->getLink()->prepare($sql);
+			$result = $resultat->execute();
+			if ($result) {
+				$query = ("DELETE FROM registered WHERE ID_Registered = '$id'");
+				$resultat = $db->getLink()->prepare($query);
+        		$resultat->execute();
+        		$proces = 1;
+			}
 		} catch(PDOException $ex) {
 			echo "An Error ocurred!";
 			some_loggging_function($ex->getMessage());
 		} finally {
-			return $resultat;
-			$_SESSION['dbconnection'] = serialize($shopDb);			
+			return $proces;			
 		}
 	}
